MaterialCache : Refactor parts into ShaderNetworkAlgo
